move src/speck/speck.h to include/speck/speck.h
