+++ src/speck/speck_common.c
@@ -2,8 +2,8 @@
 // Created by Naruto TAKAHASHI on 2017/09/02.
 //
 
-#include "speck.h"
+#include <speck/speck.h>
 #include "speck_private.h"
 
-static int foo[SPECK_ENCRYPT_TYPE_MAX];
-static int hoge[SPECK_BLOCK_CIPHER_MODE_MAX];
+//static int foo[SPECK_ENCRYPT_TYPE_MAX];
+//static int hoge[SPECK_BLOCK_CIPHER_MODE_MAX];
